Reorganise type_deduction_folder
