+++ include/api/field_descriptors.hpp
@@ -10,10 +10,10 @@
 #include "../field_size/field_size_deduce.hpp"
 #include "../field_validation/field_value_constraints.hpp"
 // #include "../field_compute/computation_from_fields.hpp"
-#include "../type_deduction/type_deduction_traits.hpp"
+#include "../type_deduction/type/type_deduction_traits.hpp"
 #include "../lib/containers/static_array.hpp"
 #include "../lib/containers/static_set.hpp"
-#include "../type_deduction/type_deduction_metafunctions.hpp"
+#include "../type_deduction/type/type_deduction_metafunctions.hpp"
 
 
 namespace s2s {
