+++ include/type_deduction/if_else_ladder/clause.hpp
@@ -1,8 +1,8 @@
-#ifndef _TYPE_DEDUCTION_CLAUSE_HPP_
-#define _TYPE_DEDUCTION_CLAUSE_HPP_
+#ifndef _CLAUSE_HPP_
+#define _CLAUSE_HPP_
 
-#include "type_deduction_tags.hpp"
-#include "../field_compute/computation_from_fields_traits.hpp"
+#include "../utils/type_tags.hpp"
+#include "../../field_compute/computation_from_fields_traits.hpp"
 
 
 namespace s2s {
@@ -17,4 +17,4 @@ struct branch {
 };
 } /* namespace s2s */
 
-#endif // _TYPE_DEDUCTION_CLAUSE_HPP_
+#endif // _CLAUSE_HPP_
