+++ include/type_deduction/if_else_ladder/ladder.hpp
@@ -1,11 +1,12 @@
-#ifndef _TYPE_DEDUCTION_LADDER_HPP_
-#define _TYPE_DEDUCTION_LADDER_HPP_
+#ifndef _LADDER_HPP_
+#define _LADDER_HPP_
 
-#include "../error/cast_error.hpp"
-#include "type_deduction_clause.hpp"
-#include "type_deduction_clause_traits.hpp"
-#include "type_deduction_helper.hpp"
+
 #include <expected>
+#include "../../error/cast_error.hpp"
+#include "clause.hpp"
+#include "clause_traits.hpp"
+#include "../utils/helper.hpp"
 
 
 namespace s2s {
@@ -17,4 +18,4 @@ struct type_if_else {
 };
 } /* namespace s2s */
 
-#endif // _TYPE_LADDER_HPP_
+#endif // _LADDER_HPP_
