+++ include/type_deduction/type/type_deduction_traits.hpp
@@ -1,10 +1,10 @@
 #ifndef _TYPE_DEDUCTION_TRAITS_HPP_
 #define _TYPE_DEDUCTION_TRAITS_HPP_
 
-#include "../field_compute/computation_from_fields.hpp"
-#include "type_deduction_switch_traits.hpp"
-#include "type_deduction_ladder_traits.hpp"
-#include "type_deduction.hpp"
+#include "../../field_compute/computation_from_fields.hpp"
+#include "../switch/switch_traits.hpp"
+#include "../if_else_ladder/ladder_traits.hpp"
+#include "type.hpp"
 
 
 namespace s2s {
