+++ include/type_deduction/utils/helper.hpp
@@ -1,12 +1,12 @@
-#ifndef _TYPE_DEDUCTION_HELPER_HPP_
-#define _TYPE_DEDUCTION_HELPER_HPP_
+#ifndef _HELPER_HPP_
+#define _HELPER_HPP_
 
 
 #include <optional>
 #include <expected>
-#include "../error/cast_error.hpp"
-#include "type_deduction_match_case_traits.hpp"
-#include "type_deduction_clause_traits.hpp"
+#include "../../error/cast_error.hpp"
+#include "../switch/match_case_traits.hpp"
+#include "../if_else_ladder/clause_traits.hpp"
 
 
 namespace s2s {
